refactor: cleanup mail encoder slightly
